+++ src/backend/TrafficCourts/Workflow.Service/Services/EmailTemplates/VerificationEmailTemplate.cs
@@ -41,7 +41,7 @@ If you need more help, contact the Violation Ticket Centre toll free 1-877-661-8
 
     private Uri CreateEmailVerificationUrl(SendEmailVerificationEmail data)
     {
-        // https://tickets.gov.bc.ca/email/verify?token={token}
+        // https://tickets.gov.bc.ca/email/verify/{token}
 
         var format = _emailConfiguration.EmailVerificationUrl;
         if (string.IsNullOrEmpty(format))
@@ -52,7 +52,7 @@ If you need more help, contact the Violation Ticket Centre toll free 1-877-661-8
         }
 
         var token = _encoder.Encode(new DisputeEmailVerificationToken { NoticeOfDisputeGuid = data.NoticeOfDisputeGuid, Token = data.Token });
-        string uri = format + "?token=" + token;
+        string uri = format + "/" + token;
 
         return new Uri(uri);
     }
