+++ src/frontend/citizen-portal/src/app/auth/interceptors/auth-http.interceptor.ts
@@ -1,17 +1,17 @@
 import { Injectable } from '@angular/core';
 import { HttpRequest, HttpHandler, HttpInterceptor, HttpResponse, HttpEvent } from '@angular/common/http';
 import { filter, Observable, of } from 'rxjs';
-import { AuthConfigService } from '../services/auth-config.service';
 import { select, Store } from '@ngrx/store';
 import { AuthStore } from '../store';
 import { AppConfigService } from 'app/services/app-config.service';
+import { AuthConfig } from '../models/auth-config.model';
 
 @Injectable({ providedIn: "root" })
 export class AuthHttpInterceptor implements HttpInterceptor {
   private token: string;
 
   constructor(
-    private authConfigService: AuthConfigService,
+    private authConfig: AuthConfig,
     private appConfigService: AppConfigService,
     private store: Store
   ) {
@@ -21,7 +21,7 @@ export class AuthHttpInterceptor implements HttpInterceptor {
   }
 
   intercept(request: HttpRequest<any>, next: HttpHandler) {
-    if (request.url === this.authConfigService.config?.authority + "/.well-known/openid-configuration") {
+    if (request.url === this.authConfig.config?.authority + "/.well-known/openid-configuration") {
       return this.handleWellKnownRequest(request.method);
     }
     if (this.token && request.url.startsWith(this.appConfigService.apiBaseUrl)) {
@@ -39,7 +39,7 @@ export class AuthHttpInterceptor implements HttpInterceptor {
       case "GET":
         return of(new HttpResponse<unknown>({
           status: 200,
-          body: this.authConfigService.authWellKnownDocument
+          body: this.authConfig.authWellKnownDocument
         }))
     }
   }
