+++ src/frontend/citizen-portal/src/app/auth/models/auth-config.model.ts
@@ -1,0 +1,6 @@
+import { OpenIdConfiguration } from "angular-auth-oidc-client";
+
+export class AuthConfig {
+  config: OpenIdConfiguration;
+  authWellKnownDocument: any;
+}
